requirements.txt
python-dotenv
pyrotgfork
tgcrypto
Pillow
aiohttp